Cleanup: Delete gRPC Server Code
Removes the implementation of gRPC server in an eclipse instance.
gRPC so far did not yet implement any analysis. Only project management and file management.
gRPC is a valid approach to create an interface that is available in 11 languages. However since we want full language agnosticity we need an additional REST server. As such having the REST server directly inside the eclipse instance reduces the indirection and there is no need to convert between JSON, proto bufs and Java objects.