Cleanup: Delete protocol buffers definitions
We used gRPC in combination with protocol buffers.
This maven project was used to store the defintiions for the protocol buffers and compile them to the respective java classes.
The compilation of proto defintions to the language specific code in the case of Java would have needed more work to create a useful build pipeline. Right now it is a Maven project just for compiling the proto bufs. Then compiled Java classes then need to be manually compied into the actual Eclipse projects to use them there.