Cleanup: Delete RMI Eclipse server side implementation
These plugin projects implement the components of our interface with RMI.
Because RMI is only useful with Java related technologies, we need an additional REST server. RMI is only used to communicate between the REST server and the eclipse instance. Having the REST server directly in Eclipse makes the RMI implementation obsolet